requirements.txt
requests ~> 2.0
numpy
pydotenv